Don't verify path traversal for performance reasons.
As mentioned in the comments genMusicSQLiteDB can probably assume
non-adversarial input.

While it's not a terrible idea to do additional checks, there is a
notable performance penalty and overall the goal of the software is
geared more towards speed.

Additionally, the primary reason the check was implemented was to make
Gosec satisfied, but Gosec appears to not properly detect the fix
(or perhaps the fix is insufficient regardless).