Apply nosec G202 for PrepareStatementRemove generator.
Reason: For intended use case (famous last words), tables and column
names are considered "Trusted", and we consider the real-world risk of
SQLi as minimal in these inputs for the sake of practicality.

HOWEVER, the content of the table and particular the input in the rows
should take advantage of properly formed prepared statements to avoid
issues with SQLi.